refactor to-be-shared methods out of differential test
